backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
bcrypt
python-jose
passlib
collections
sklearn
pandas
numpy
matplotlib
